backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Core API
fastapi>=0.115
uvicorn>=0.30
python-dotenv>=1.0.0

# AI & Vector DB
torch
groq>=0.11
pinecone==8.1.0
fastembed>=0.4.0

# Framework
langchain>=0.3
langchain-community>=0.3
langchain-groq>=0.2